+++ include/ironbee/server.h
@@ -22,7 +22,9 @@
  * @file
  * @brief IronBee --- Ironbee as a server plugin
  *
- * @author Brian Rectanus <brectanus@qualys.com>, Nick Kew <nkew@qualys.com>
+ * @author Brian Rectanus <brectanus@qualys.com>
+ * @author Nick Kew <nkew@qualys.com>
+ * @author Christopher Alfeld <calfeld@qualys.com>
  */
 
 #include <ironbee/engine_types.h>
@@ -89,23 +91,29 @@ typedef enum {
  */
 typedef ib_status_t (*ib_server_error_fn_t)(
     ib_tx_t *tx,
-    int status,
-    void *cbdata
+    int      status,
+    void    *cbdata
 );
+
 /**
  * Set server error header.
  *
  * @param[in] tx The transaction.
- * @param[in] name The null-terminated name of the header.
- * @param[in] name The null-terminated value of the header.
+ * @param[in] name Name of header.
+ * @param[in] name_length Length of @a name.
+ * @param[in] value value of header.
+ * @param[in] value_length Length of @a value.
  * @param[in] cbdata Callback data.
  */
 typedef ib_status_t (*ib_server_error_hdr_fn_t)(
-    ib_tx_t *tx,
+    ib_tx_t    *tx,
     const char *name,
+    size_t      name_length,
     const char *value,
-    void *cbdata
+    size_t      value_length,
+    void       *cbdata
 );
+
 /**
  * Set server error data.
  *
@@ -115,10 +123,10 @@ typedef ib_status_t (*ib_server_error_hdr_fn_t)(
  * @param[in] cbdata Callback data.
  */
 typedef ib_status_t (*ib_server_error_data_fn_t)(
-    ib_tx_t *tx,
-    const uint8_t *data,
-    size_t dlen,
-    void *cbdata
+    ib_tx_t    *tx,
+    const char *data,
+    size_t      dlen,
+    void       *cbdata
 );
 
 /**
@@ -127,34 +135,24 @@ typedef ib_status_t (*ib_server_error_data_fn_t)(
  * @param[in] tx The transaction.
  * @param[in] dir The direction.
  * @param[in] action The action determining how to add the header.
- * @param[in] hdr The header name. A null-terminated string.
- * @param[in] value The header value. A null-terminated string.
+ * @param[in] name Name of header.
+ * @param[in] name_length Length of @a name.
+ * @param[in] value value of header.
+ * @param[in] value_length Length of @a value.
  * @param[in] rx The regular expression if the action is @ref IB_HDR_EDIT.
  * @param[in] cbdata Callback data.
  */
 typedef ib_status_t (*ib_server_header_fn_t)(
-    ib_tx_t *tx,
-    ib_server_direction_t dir,
-    ib_server_header_action_t action,
-    const char *hdr,
-    const char *value,
-    ib_rx_t *rx,
-    void *cbdata
+    ib_tx_t                   *tx,
+    ib_server_direction_t      dir,
+    ib_server_header_action_t  action,
+    const char                *name,
+    size_t                     name_length,
+    const char                *value,
+    size_t                     value_length,
+    ib_rx_t                   *rx,
+    void                      *cbdata
 );
-
-#ifdef HAVE_FILTER_DATA_API
-typedef ib_status_t (*ib_server_filter_init_fn_t)(
-    ib_tx_t *tx,
-    ib_server_direction_t dir,
-    void *cbdata
-);
-typedef ib_status_t (*ib_server_filter_data_fn_t)(
-    ib_tx_t *tx,
-    ib_server_direction_t dir,
-    const char *block, size_t len,
-    void *cbdata
-);
-#endif /* HAVE_FILTER_DATA_API */
 
 /**
  * Close the given connection.
@@ -172,9 +170,24 @@ typedef ib_status_t (*ib_server_filter_data_fn_t)(
  */
 typedef ib_status_t (*ib_server_close_fn_t)(
     ib_conn_t *conn,
-    ib_tx_t *tx,
-    void *cbdata
+    ib_tx_t   *tx,
+    void      *cbdata
 );
+
+#ifdef HAVE_FILTER_DATA_API
+typedef ib_status_t (*ib_server_filter_init_fn_t)(
+    ib_tx_t               *tx,
+    ib_server_direction_t  dir,
+    void                  *cbdata
+);
+typedef ib_status_t (*ib_server_filter_data_fn_t)(
+    ib_tx_t               *tx,
+    ib_server_direction_t  dir,
+    const char            *block,
+    size_t                 len,
+    void                  *cbdata
+);
+#endif /* HAVE_FILTER_DATA_API */
 
 struct ib_server_t {
     /* Header */
@@ -216,6 +229,12 @@ struct ib_server_t {
     /** Callback data for err_body_fn */
     void *err_body_data;
 
+    /** Close connection. */
+    ib_server_close_fn_t close_fn;
+
+    /** Callback data for close_fn. */
+    void *close_data;
+
 #ifdef HAVE_FILTER_DATA_API
     /** Initialize data filtering */
     ib_server_filter_init_fn_t init_fn;
@@ -228,17 +247,9 @@ struct ib_server_t {
 
     /** Callback data for data_fn */
     void *data_data;
-
 #endif
-
-    /** Close connection. */
-    ib_server_close_fn_t close_fn;
-
-    /** Callback data for close_fn. */
-    void *close_data;
 };
 
-#ifdef DOXYGEN
 /**
  * Function to indicate an error.
  * Status argument is an HTTP response code, or a special value
@@ -249,15 +260,15 @@ struct ib_server_t {
  * In the second instance, the server takes an enumerated special
  * action, or returns NOTIMPL if that's not supported.
  *
- * @param[in] svr Server object
+ * @param[in] svr The @ref ib_server_t
  * @param[in] tx Transaction
  * @param[in] status Status code
  * @return indication of whether the requested error action is supported
  */
-ib_status_t ib_server_error_response(
-    ib_server_t *svr,
-    ib_tx_t     *tx,
-    int          status
+ib_status_t DLL_PUBLIC ib_server_error_response(
+    const ib_server_t *svr,
+    ib_tx_t           *tx,
+    int                status
 );
 
 /**
@@ -265,17 +276,21 @@ ib_status_t ib_server_error_response(
  * Any values set here will only take effect if an HTTP response
  * code is also set using ib_server_error_response.
  *
- * @param[in] svr Server object
+ * @param[in] svr The @ref ib_server_t
  * @param[in] tx Transaction object
- * @param[in] hdr Header to set
- * @param[in] value Value to set header to.
+ * @param[in] name Name of header.
+ * @param[in] name_length Length of @a name.
+ * @param[in] value value of header.
+ * @param[in] value_length Length of @a value.
  * @return indication of whether the requested error action is supported
  */
-ib_status_t ib_server_error_header(
-    ib_server_t *svr,
-    ib_tx_t     *tx,
-    const char  *hdr,
-    const char  *value
+ib_status_t DLL_PUBLIC ib_server_error_header(
+    const ib_server_t *svr,
+    ib_tx_t           *tx,
+    const char        *name,
+    size_t             name_length,
+    const char        *value,
+    size_t             value_length
 );
 
 /**
@@ -283,16 +298,17 @@ ib_status_t ib_server_error_header(
  * Any values set here will only take effect if an HTTP response
  * code is also set using ib_server_error_response.
  *
- * @param[in] svr The ib_server_t
+ * @param[in] svr The @ref ib_server_t
  * @param[in] ctx Application pointer from the server
  * @param[in] data Response to set
- * @param[in] cbdata Callback data.
+ * @param[in] dlen Length of @a data
  * @return indication of whether the requested error action is supported
  */
-ib_status_t ib_server_error_header(
-    ib_server_t *svr,
-    ib_tx_t     *tx,
-    const char  *data
+ib_status_t DLL_PUBLIC ib_server_error_body(
+    const ib_server_t *svr,
+    ib_tx_t           *tx,
+    const char        *data,
+    size_t             dlen
 );
 
 /**
@@ -305,24 +321,47 @@ ib_status_t ib_server_error_header(
  * We exclude the "edit" option on the premise that Ironbee will perform
  * any such operation internally and use set/append/merge/add/unset
  *
- * @param[in] svr The ib_server_t
- * @param[in] ctx Application pointer from the server
+ * @param[in] svr The @ref ib_server_t
+ * @param[in] tx Transaction
  * @param[in] dir Request or Response
  * @param[in] action Action requested
- * @param[in] hdr Header to act on
- * @param[in] value Value to act with
- * @param[in] cbdata Callback data.
- * @return The action actually performed, or an error
+ * @param[in] name Name of header.
+ * @param[in] name_length Length of @a name.
+ * @param[in] value value of header.
+ * @param[in] value_length Length of @a value.
+ * @param[in] rx The regular expression if the action is @ref IB_HDR_EDIT.
+ * @return Success or error.
  */
-ib_server_header_action_t ib_server_header(
-    ib_server_t               *svr,
-    void                      *ctx,
+ib_status_t DLL_PUBLIC ib_server_header(
+    const ib_server_t         *svr,
     ib_tx_t                   *tx,
     ib_server_direction_t      dir,
     ib_server_header_action_t  action,
-    const char                *hdr,
+    const char                *name,
+    size_t                     name_length,
     const char                *value,
-    void                      *cbdata
+    size_t                     value_length,
+    ib_rx_t                   *rx
+);
+
+/**
+ * Close the given connection.
+ *
+ * @param[in] svr The @ref ib_server_t
+ * @param[in] conn Connection to close.
+ * @param[in] tx Transaction that the connection is in.
+ *               There are situations where tx may be NULL.
+ *               Implementers should expect to close a
+ *               connection when there is no associated transaction.
+ *
+ * @returns
+ *   - IB_OK on success.
+ *   - IB_DECLINED if the server cannot honor this request.
+ */
+ib_status_t DLL_PUBLIC ib_server_close(
+    const ib_server_t *svr,
+    ib_conn_t         *conn,
+    ib_tx_t           *tx
 );
 
 #ifdef HAVE_FILTER_DATA_API
@@ -336,72 +375,38 @@ ib_server_header_action_t ib_server_header(
  * consuming its entire input, and generating the entire payload as
  * output in blocks.
  *
- * @param[in] svr The ib_server_t
+ * @param[in] svr The @ref ib_server_t
  * @param[in] ctx Application pointer from the server
  * @param[in] dir Request or Response
- * @param[in] cbdata Callback data.
  * @return Indication of whether the action is supported and will happen.
  */
-ib_status_t ib_server_filter_init(
-    ib_server_t    *svr,
-    void           *ctx,
-    ib_direction_t  dir,
-    void           *cbdata
+ib_status_t DLL_PUBLIC ib_server_filter_init(
+    const ib_server_t *svr,
+    void              *ctx,
+    ib_direction_t     dir
 );
 
 /**
  * Filtered data should only be passed if ib_server_filter_init returned IB_OK.
  *
- * @param[in] svr The ib_server_t
+ * @param[in] svr The @ref ib_server_t
  * @param[in] ctx Application pointer from the server
  * @param[in] dir Request or Response
  * @param[in] data Data chunk
  * @param[in] len Length of chunk
- * @param[in] cbdata Callback data.
  * @return Success or error
  */
-ib_status_t ib_server_filter_data(
-    ib_server_t    *svr,
-    void           *ctx,
-    ib_direction_t  dir,
-    const char     *block,
-    size_t          len,
-    void           *cbdata
+ib_status_t DLL_PUBLIC ib_server_filter_data(
+    const ib_server_t *svr,
+    void              *ctx,
+    ib_direction_t     dir,
+    const char        *block,
+    size_t             len
 );
 #endif /* HAVE_FILTER_DATA_API */
 
-#else /* DOXYGEN */
-
-#define ib_server_error_response(svr, tx, status) \
-    ((svr) && (svr)->err_fn) ? (svr)->err_fn(tx, status, (svr)->err_data) \
-                  : IB_ENOTIMPL
-#define ib_server_error_header(svr, tx, name, val) \
-    ((svr) && (svr)->err_hdr_fn) ? (svr)->err_hdr_fn(tx, name, val, (svr)->err_hdr_data) \
-                      : IB_ENOTIMPL
-#define ib_server_error_body(svr, tx, data, dlen) \
-    ((svr) && (svr)->err_body_fn) ? (svr)->err_body_fn(tx, data, dlen, (svr)->err_body_data) \
-                       : IB_ENOTIMPL
-#define ib_server_header(svr, tx, dir, action, hdr, value, rx) \
-    ((svr) && (svr)->hdr_fn) ? (svr)->hdr_fn(tx, dir, action, hdr, value, (svr)->hdr_data, rx) \
-                  : IB_ENOTIMPL
-
-#ifdef HAVE_FILTER_DATA_API
-#define ib_server_filter_init(svr, tx, dir) \
-    ((svr) && (svr)->init_fn) ? (svr)->init_fn(tx, dir, (svr)->init_data) \
-                   : IB_ENOTIMPL
-#define ib_server_filter_data(svr, tx, dir, data, len) \
-    ((svr) && (svr)->data_fn) ? (svr)->data_fn(tx, dir, data, len, (svr)->data_data) \
-                   : IB_ENOTIMPL
-#endif /* HAVE_FILTER_DATA_API */
-
-#define ib_server_error_close(svr, conn, tx) \
-    ((svr) && (svr)->close_fn) ? (svr)->close_fn(conn, tx, (svr)->close_data) \
-                  : IB_ENOTIMPL
-
-#endif /* DOXYGEN */
-
 /**
- * @} IronBeePlugins
+ * @} IronBeeServer
  */
 
 #ifdef __cplusplus