+++ modules/moddevel_txresp.c
@@ -105,14 +105,15 @@ static ib_status_t tx_header_finished(
 
     /* Note: ib_server_header() ignores lengths for now */
     rc = ib_tx_server_header(tx, IB_SERVER_RESPONSE, IB_HDR_SET,
-                             "ENGINE-UUID", ib_engine_instance_uuid_str(ib),
+                             IB_S2SL("ENGINE-UUID"),
+                             IB_S2SL(ib_engine_instance_uuid_str(ib)),
                              NULL);
     if (rc != IB_OK) {
         return rc;
     }
 
     rc = ib_tx_server_header(tx, IB_SERVER_RESPONSE, IB_HDR_SET,
-                             "TX-UUID", tx->id, NULL);
+                             IB_S2SL("TX-UUID"), IB_S2SL(tx->id), NULL);
     if (rc != IB_OK) {
         return rc;
     }